Add "Diagnoses Starter Kit" with 1k+ CIEL Diagnoses with French (#84)
This new collection contains all the available codes in CIEL as of July 16 2025 that are BOTH: Diagnosis codes, AND contain translations for at least French. This yields a collection of over 1,400 diagnosis concepts which should give implementations a reasonable Starter-Kit of diagnoses. Since CIEL contains so many mappings to other standards like ICD and SNOMED, this can be used as a reasonable starting place for many implementers needing to experiment with OpenMRS.

Since we do not have time/scope in this phase of the DRC EMR MVP to add ICD-11 or ICD-10 support in French, this should unblock the need for more codes in the MVP.