Update Diagnoses concepts for total of 2k+ diagnoses (#87)
* Update Diagnoses concepts for total of 2k+ diagnoses

About: This PR adds version 2 of the Diagnoses Starter Kit, which has >2,100 diagnosis concepts (all with french translations) as opposed to version 1 which only had about 1,400. 
Context: I re-did the "add to collection" workflow from CIEL and this seems to have brought in the concepts that were missed in version 1. 

Also deletes the previous version 1.

* Delete configuration/backend_configuration/ocl/openmrs_DiagnosesStarterKit_v1_41691.2025-07-16_045128.zip